feat(api): add translation script and reorgenise all process in 3 differents scripts + add folder to feed the front with .srt and .json exports
