+++ .gitattributes
@@ -1,1 +1,2 @@
 *.h5 filter=lfs diff=lfs merge=lfs -text
+*.wav filter=lfs diff=lfs merge=lfs -text
