+++ API/utils/stt_functions.py
@@ -3,18 +3,19 @@ import os
 from pydub import AudioSegment
 import speech_recognition as sr
 
-debug_mode = True
+debug_mode = None
 def log(message):
     if debug_mode:
         print(message)
 
 
-def process_stt():
+def process_stt(): # for whisper
     pass
 
 def process_stt_deprecated(audio_file_path, chunk_length_ms=4000):
     """Process the audio file and return a list of transcribed sentences with timestamps"""
     print(f"Processing audio file: {audio_file_path}")
+    print(f"Processing in progress...")
 
     chunks_dir = os.path.join(os.getcwd(), "chunks")
     create_chunks_directory(chunks_dir)
@@ -50,7 +51,7 @@ def process_stt_deprecated(audio_file_path, chunk_length_ms=4000):
                 }
                 results.append(result)
 
-            print(f"{i}. {text}")
+            log(f"{i}. {text}")
 
         os.remove(chunk_filename) # remove the chunk file as progresses
